requirements.txt
fastapi[all]
mysql-connector-python
passlib[bcrypt]
python-jose[cryptography]